created component for all todos, working on logic for sort button
