+++ client/src/pages/Todo.tsx
@@ -1,15 +1,13 @@
 import { useState } from "react";
-import { Spacer, Flex, Box, Text, useBreakpointValue, Button, } from "@chakra-ui/react";
-import axios from "axios";
-import { Card, CardHeader, CardBody, Stack} from '@chakra-ui/react'
+import { Box, Text, useBreakpointValue, } from "@chakra-ui/react";
+import { Stack} from '@chakra-ui/react'
 import React from "react";
 import TodoModal from "../components/Todo/TodoModal";
 import { useLoaderData } from "react-router-dom";
-import EditTodoModal from "../components/Todo/EditTodoModal";
-import { DeleteIcon } from "@chakra-ui/icons";
 // commenting out checkbox until working again
 import CompletedCheckBox from "../components/Todo/CompletedCheckBox";
 import SortByMenu from "../components/Todo/SortByMenu";
+import AllTodos from "../components/Todo/AllTodos";
 
 
 export type todosObject = {
@@ -31,26 +29,8 @@ const Todo = () => {
   const loadedData = data.todos.data;
   const [todoData, setTodoData] = useState(loadedData)
   const headerMargin = useBreakpointValue({ base: '1.5em', sm: '1em', md: '2em', lg: '3em'});
-  const style = {
-    border: {
-      'border': 'solid 4px #371236'
-    },
-  };
 
-  const deleteTodo = async (id: number) => {
-    try {
-      const response = await axios.delete(`http://localhost:3001/todo/${id}`);
-      return response.data;
-    } catch (error) {
-      console.error(error);
-    }
-  };
 
-  const deleteHandler = async (id: number) => {
-    deleteTodo(id);
-    setTodoData(loadedData.filter((todoData: todosObject) => todoData.id !== id));
-    window.location.reload();
-  };
 
 // when checkbox is clicked, task is 
 // placed in new array, shown when toggled
@@ -67,7 +47,8 @@ completedTodos.forEach(item => {
   }
 });
 loadedData.push(...completedTodos);
-
+// instead of creating new component for completed todos, only loop through completed todos
+// make separate loops for all sorting elements
 console.log('COMPLETED TODOS ARRAY', completedTodos);
 console.log('TODO DATA', todoData);
   return (
@@ -80,68 +61,17 @@ console.log('TODO DATA', todoData);
         mb='2em'>{date}</Text>
 
       <TodoModal setCurrentData={setTodoData}/>
-< SortByMenu />
+      < SortByMenu />
       <Stack
         className="rendered-todos"
         spacing='6'
         mx='10'>
+
         {
-          loadedData.map((todos: todosObject) => {
+          todoData.map((todos: todosObject) => {
             return (
-              <Card
-                overflow='hidden'
-                backgroundColor='#FFFFFA'
-                style={style.border}
-                key={todos?.id}
-                size={{base: 'xs', sm: 'xs', md: 'sm', lg: 'sm'}}
-                color='black'
-                pb='1em'>
-                <CardHeader
-                  className='card-header'
-                  display='flex'
-                  alignContent='flex-end'
-                  fontSize='lg'
-                  pt={{base: '1em', sm: '1em'}}>
-                   <CompletedCheckBox 
-                   id={todos?.id} 
-                   setData={setTodoData}/>
-                  <Text
-                    fontWeight='800'
-                  > {todos?.todaysDate}</Text>
-                  <Spacer></Spacer>
-                  <Text
-                  className='priority'
-                    color={todos?.priority === 'Extremely Important!' ? "red" : todos?.priority === 'Coming soon.' ? "orange" : "green"}
-                    mr='1em'
-                    fontWeight='700'>{todos?.priority}</Text>
-                </CardHeader>
-                <CardBody
-                  fontSize='md'>
-                  <Text my='10px'>{todos?.todo}</Text>
-                  <Text>
-                    {todos?.reflectionText}
-                  </Text>
-                
-                </CardBody>
-                <Flex
-          
-                  gap='3'
-                  flexDirection='row'
-                  justify={{base: 'center', sm: 'center', md:'flex-end', lg: 'flex-end'}}
-                  mr={{base: '0', sm: '0', md: '2em', lg: '2em'}}
-                >
-                   <EditTodoModal id={todos?.id}/>
-                    <DeleteIcon 
-                    boxSize={6}
-                    color='#371236'
-                    onClick={() => deleteHandler(todos?.id)}
-                    cursor='pointer'
-                 
-                    />
-                   
-                </Flex>
-                
-              </Card>
+              <AllTodos id={todos?.id} priority={todos?.priority} todo={todos?.todo} reflectionText={todos?.reflectionText} todaysDate={todos?.todaysDate} 
+              />
             )
           })
         }
